CLI-837 Move tar.ts from core/host/install to core/io
Tar reading is generic file-format parsing with no host/install
coupling, so it belongs in core/io alongside other I/O primitives.
Update CLAUDE.md reference accordingly.
